move fixtures, very basic test for sync
