+++ python/tests/test_check.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 from itertools import chain
-from pathlib import Path
 from unittest.mock import patch
 
 import pytest
@@ -9,87 +8,13 @@ import pytest
 from tach.check import check_import
 from tach.cli import tach_check
 from tach.core import (
-    Dependency,
     ModuleConfig,
-    ModuleNode,
-    ModuleTree,
 )
 from tach.core.config import RootModuleConfig
 from tach.filesystem import validate_project_modules
 
-
-@pytest.fixture
-def example_dir() -> Path:
-    current_dir = Path(__file__).parent
-    return current_dir / "example"
-
-
-@pytest.fixture
-def test_config() -> ModuleConfig:
-    return ModuleConfig(path="test", strict=False)
-
-
-@pytest.fixture
-def module_tree() -> ModuleTree:
-    return ModuleTree(
-        root=ModuleNode(
-            is_end_of_path=False,
-            full_path="",
-            config=None,
-            children={
-                "domain_one": ModuleNode(
-                    is_end_of_path=True,
-                    full_path="domain_one",
-                    config=ModuleConfig(
-                        path="domain_one",
-                        depends_on=[
-                            Dependency(path="domain_one.subdomain", deprecated=True),
-                            Dependency(path="domain_three"),
-                        ],
-                        strict=True,
-                    ),
-                    interface_members=["public_fn"],
-                    children={
-                        "subdomain": ModuleNode(
-                            is_end_of_path=True,
-                            full_path="domain_one.subdomain",
-                            config=ModuleConfig(
-                                path="domain_one.subdomain", strict=True
-                            ),
-                            children={},
-                        )
-                    },
-                ),
-                "domain_two": ModuleNode(
-                    is_end_of_path=True,
-                    full_path="domain_two",
-                    config=ModuleConfig(
-                        path="domain_two",
-                        depends_on=[Dependency(path="domain_one")],
-                        strict=False,
-                    ),
-                    children={
-                        "subdomain": ModuleNode(
-                            is_end_of_path=True,
-                            full_path="domain_two.subdomain",
-                            config=ModuleConfig(
-                                path="domain_two",
-                                depends_on=[Dependency(path="domain_one")],
-                                strict=False,
-                            ),
-                            children={},
-                        )
-                    },
-                ),
-                "domain_three": ModuleNode(
-                    is_end_of_path=True,
-                    full_path="domain_three",
-                    config=ModuleConfig(path="domain_three", strict=False),
-                    children={},
-                ),
-            },
-        )
-    )
+# from tests.conftest import *  # noqa
+# from tests.conftest import example_dir  # noqa
 
 
 @pytest.mark.parametrize(
@@ -164,11 +89,14 @@ def test_check_deprecated_import(module_tree):
     assert check_error.is_deprecated
 
 
-def test_valid_example_dir(example_dir):
+def test_valid_example_dir(example_dir, capfd):
     project_root = example_dir / "valid"
     with pytest.raises(SystemExit) as exc_info:
         tach_check(project_root=project_root)
     assert exc_info.value.code == 0
+    captured = capfd.readouterr()
+    assert "✅" in captured.out  # success state
+    assert "‼️" in captured.err  # deprecated warning
 
 
 def test_valid_example_dir_monorepo(example_dir):
